python_service/requirements.txt
chardet==4.0.0
click==7.1.2
google-api-python-client==2.65.0
google-auth-oauthlib==0.7.1
ipfshttpclient==0.8.0a1
MarkupSafe==1.1.1
aniso8601==8.1.0
blinker==1.4
certifi==2020.12.5
Flask==1.1.2
flask_cors==3.0.10
Flask-Jsonpify==1.5.0
flask-requests==0.0.14
Flask-RESTful==0.3.8
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
pampy==0.3.0
python-dotenv==1.0.0
pytz==2020.4
requests==2.26.0
six==1.15.0
urllib3==1.26.3
Werkzeug==1.0.1